factor out machine-dependent definitions and helpers; add ARCH, ARCHDIR and CROSS_COMPILE to build; consolidate natively-built tools under common dir
