+++ .circleci/config.yml
@@ -40,7 +40,7 @@ jobs:
       - run: sudo gcloud config set project ${GOOGLE_PROJECT_ID}
       - run: gcloud --quiet config set compute/zone ${GOOGLE_COMPUTE_ZONE}
 
-      - run: mkdir temp && cp output/stage3/bin/stage3.img temp/ && cp output/boot/boot.img temp/ && cp output/mkfs/bin/mkfs temp/
+      - run: mkdir temp && cp output/stage3/bin/stage3.img temp/ && cp output/boot/boot.img temp/ && cp output/tools/bin/mkfs temp/
       - run: cd temp && tar cvzf nanos-nightly-linux.tar.gz * && gsutil cp nanos-nightly-linux.tar.gz gs://nanos/release/nightly
       - run: gsutil acl ch -u AllUsers:R gs://nanos/release/nightly/nanos-nightly-linux.tar.gz
       - run: rm -r temp
@@ -92,7 +92,7 @@ jobs:
       - run: sudo gcloud config set project ${GOOGLE_PROJECT_ID}
       - run: gcloud --quiet config set compute/zone ${GOOGLE_COMPUTE_ZONE}
 
-      - run: mkdir temp && cp output/stage3/bin/stage3.img temp/ && cp output/boot/boot.img temp/ && cp output/mkfs/bin/mkfs temp/
+      - run: mkdir temp && cp output/stage3/bin/stage3.img temp/ && cp output/boot/boot.img temp/ && cp output/tools/bin/mkfs temp/
       - run: cd temp && tar cvzf nanos-nightly-darwin.tar.gz * && gsutil cp nanos-nightly-darwin.tar.gz gs://nanos/release/nightly
       - run: gsutil acl ch -u AllUsers:R gs://nanos/release/nightly/nanos-nightly-darwin.tar.gz
       - run: rm -r temp
