+++ README.md
@@ -112,7 +112,7 @@ develop locally. If that's the case a commonly used idiom is to simply
 copy the 3 required files to an appropriate release:
 
 ```
-cp output/mkfs/bin/mkfs ~/.ops/0.1.17/.
+cp output/tools/bin/mkfs ~/.ops/0.1.17/.
 cp output/boot/boot.img ~/.ops/0.1.17/.
 cp output/stage3/bin/stage3.img ~/.ops/0.1.17/.
 ```
