--- contgen/Makefile
@@ -1,6 +1,0 @@
-PROGRAMS=	contgen
-SRCS-contgen=	$(CURDIR)/contgen.c
-
-all: $(PROGRAMS)
-
-include ../rules.mk
