+++ src/http/http.c
@@ -1,4 +1,4 @@
-#include <kernel.h>
+#include <runtime.h>
 #include <http.h>
 
 #define STATE_INIT 0
