+++ src/kernel/backed_heap.c
