+++ src/kernel/elf.c
