+++ src/kernel/kernel.c
