+++ src/kernel/kvm_platform.c
