+++ src/kernel/kvm_platform.h
