+++ src/kernel/symtab.c
