+++ src/kernel/symtab.h
