+++ src/kernel/vdso-now.c
