+++ src/kernel/vdso.c
