+++ src/kernel/xen_platform.h
