+++ stage3/Makefile
@@ -12,6 +12,17 @@ SRCS-stage3.img= \
 	$(SRCDIR)/gdb/gdbtcp.c \
 	$(SRCDIR)/gdb/gdbutil.c \
 	$(SRCDIR)/http/http.c \
+	$(SRCDIR)/kernel/backed_heap.c \
+	$(SRCDIR)/kernel/elf.c \
+	$(SRCDIR)/kernel/kernel.c \
+	$(SRCDIR)/kernel/kvm_platform.c \
+	$(SRCDIR)/kernel/pagecache.c \
+	$(SRCDIR)/kernel/pci.c \
+	$(SRCDIR)/kernel/pvclock.c \
+	$(SRCDIR)/kernel/schedule.c \
+	$(SRCDIR)/kernel/service.c \
+	$(SRCDIR)/kernel/symtab.c \
+	$(SRCDIR)/kernel/vdso-now.c \
 	$(SRCDIR)/net/direct.c \
 	$(SRCDIR)/net/net.c \
 	$(SRCDIR)/net/netsyscall.c \
@@ -86,28 +97,17 @@ SRCS-stage3.img= \
 	$(SRCDIR)/hyperv/storvsc/storvsc.c \
 	$(SRCDIR)/x86_64/ap.s \
 	$(SRCDIR)/x86_64/apic.c \
-	$(SRCDIR)/x86_64/backed_heap.c \
 	$(SRCDIR)/x86_64/breakpoint.c \
 	$(SRCDIR)/x86_64/clock.c \
 	$(SRCDIR)/x86_64/crt0.s \
-	$(SRCDIR)/x86_64/elf.c \
 	$(SRCDIR)/x86_64/flush.c \
 	$(SRCDIR)/x86_64/hpet.c \
 	$(SRCDIR)/x86_64/interrupt.c \
-	$(SRCDIR)/x86_64/kernel.c \
-	$(SRCDIR)/x86_64/kvm_platform.c \
 	$(SRCDIR)/x86_64/mp.c \
 	$(SRCDIR)/x86_64/page.c \
-	$(SRCDIR)/x86_64/pagecache.c \
-	$(SRCDIR)/x86_64/pci.c \
-	$(SRCDIR)/x86_64/pvclock.c \
 	$(SRCDIR)/x86_64/rtc.c \
-	$(SRCDIR)/x86_64/schedule.c \
 	$(SRCDIR)/x86_64/serial.c \
-	$(SRCDIR)/x86_64/service.c \
-	$(SRCDIR)/x86_64/symtab.c \
 	$(SRCDIR)/x86_64/synth.c \
-	$(SRCDIR)/x86_64/vdso-now.c \
 	$(SRCDIR)/x86_64/x2apic.c \
 	$(SRCDIR)/x86_64/xapic.c \
 	$(VDSO_OBJDIR)/vdso-image.c \
@@ -144,21 +144,22 @@ SRCS-lwip= \
 
 INCLUDES=\
 	-I$(SRCDIR) \
+	-I$(ARCHDIR) \
 	-I$(SRCDIR)/gdb \
 	-I$(SRCDIR)/http \
+	-I$(SRCDIR)/kernel \
 	-I$(SRCDIR)/net \
 	-I$(SRCDIR)/runtime \
 	-I$(SRCDIR)/tfs \
 	-I$(SRCDIR)/unix \
 	-I$(SRCDIR)/unix_process \
-	-I$(SRCDIR)/x86_64 \
 	-I$(SRCDIR)/xen/public \
 	-I$(SRCDIR)/hyperv/include \
 	-I$(LWIPDIR)/src/include
 
-AFLAGS+=-I$(SRCDIR)/x86_64
+AFLAGS+=-I$(ARCHDIR)
 
-CFLAGS+=$(KERNCFLAGS) -DSTAGE3 -mcmodel=kernel
+CFLAGS+=$(KERNCFLAGS) -DSTAGE3 -DKERNEL -mcmodel=kernel
 CFLAGS+=-Wno-address # lwIP build sadness
 CFLAGS+=$(INCLUDES)
 
@@ -167,32 +168,29 @@ ifeq ($(TRACE),ftrace)
 CFLAGS+= -DCONFIG_FTRACE -pg
 SRCS-stage3.img+= \
 	$(SRCDIR)/unix/ftrace.c \
-	$(SRCDIR)/x86_64/ftrace.s
+	$(ARCHDIR)/ftrace.s
 endif
 
 #CFLAGS+=	-DLWIPDIR_DEBUG -DEPOLL_DEBUG -DNETSYSCALL_DEBUG -DKERNEL_DEBUG
 AFLAGS+=	-felf64 -I$(OBJDIR)/
 LDFLAGS+=	$(KERNLDFLAGS) --undefined=_start -T linker_script
 
-VDSO_SRCDIR=    $(SRCDIR)/x86_64
+VDSOGEN=	$(TOOLDIR)/vdsogen
+VDSO_SRCDIR=    $(SRCDIR)/kernel
 VDSO_OBJDIR=    $(OBJDIR)/vdso
 VDSO_SRCS=      $(VDSO_SRCDIR)/vdso.c $(VDSO_SRCDIR)/vdso-now.c
-VDSO_OBJS=      $(patsubst $(ROOTDIR)/%.c,$(VDSO_OBJDIR)/%.o,$(VDSO_SRCS))
-VDSO_CFLAGS=    -DBUILD_VDSO -I$(INCLUDES) -I$(OBJDIR) -fPIC -c
-VDSO_LDFLAGS=   -nostdlib -fPIC -shared --build-id=none --hash-style=both --eh-frame-hdr -T$(VDSO_SRCDIR)/vdso.lds
-VDSOGEN_SRC=    $(SRCDIR)/x86_64/vdsogen.c
+VDSO_OBJS=      $(patsubst $(VDSO_SRCDIR)/%.c,$(VDSO_OBJDIR)/%.o,$(VDSO_SRCS))
+VDSO_CFLAGS=    -DKERNEL -DBUILD_VDSO -I$(INCLUDES) -I$(OBJDIR) -fPIC -c
+VDSO_LDFLAGS=   -nostdlib -fPIC -shared --build-id=none --hash-style=both --eh-frame-hdr -T$(ARCHDIR)/vdso.lds
 VDSO_DEPS=      $(patsubst %.o,%.d,$(VDSO_OBJS))
 OBJDUMPFLAGS=	-d -S -M intel-mnemonic
 
 DEPFILES+=      $(VDSO_DEPS)
-CLEANFILES+=    $(foreach f,gitversion.c frame.inc stage3.dis stage3.img src/unix/ftrace.* src/x86_64/ftrace.*,$(OBJDIR)/$f) $(VDSO_OBJDIR)/vdsogen $(VDSO_OBJDIR)/vdso.so $(VDSO_OBJDIR)/vdso-image.c $(VDSO_OBJS) $(VDSO_DEPS)
-CLEANDIRS+=     $(foreach d,output src src/hyperv vdso/src/x86_64 vdso/src vdso vendor vendor/lwip vendor/lwip/src,$(OBJDIR)/$d)
-
-msg_vdsogen_cc= CC	$@
-cmd_vdsogen_cc= $(CC) -o $@ $^
+CLEANFILES+=    $(foreach f,gitversion.c frame.inc stage3.dis stage3.img src/unix/ftrace.* $(ARCHDIR)/ftrace.*,$(OBJDIR)/$f) $(VDSO_OBJDIR)/vdso.so $(VDSO_OBJDIR)/vdso-image.c $(VDSO_OBJS) $(VDSO_DEPS)
+CLEANDIRS+=     $(foreach d,output src src/hyperv vdso/src/$(ARCH) vdso/src vdso vendor vendor/lwip vendor/lwip/src,$(OBJDIR)/$d)
 
 msg_vdsogen=    VDSOGEN	$@
-cmd_vdsogen=    $(VDSO_OBJDIR)/vdsogen $(VDSO_OBJDIR)/vdso.so $@
+cmd_vdsogen=    $(VDSOGEN) $(VDSO_OBJDIR)/vdso.so $@
 
 msg_vdso_cc=    CC	$@
 cmd_vdso_cc=    $(CC) $(DEPFLAGS) $(VDSO_CFLAGS) -c $< -o $@
@@ -217,21 +215,20 @@ include ../rules.mk
 
 stage3.dis: $(OBJDIR)/stage3.dis
 
+$(VDSOGEN):
+	@$(MAKE) -C $(ROOTDIR)/tools vdsogen
+
 $(OBJDIR)/stage3.dis: $(PROG-stage3.img)
 	$(call cmd,objdump)
 
-$(VDSO_OBJDIR)/vdsogen: $(VDSOGEN_SRC)
-	@$(MKDIR) $(VDSO_OBJDIR)
-	$(call cmd,vdsogen_cc)
-
-$(VDSO_OBJDIR)/src/x86_64/%.o: $(VDSO_SRCDIR)/%.c
+$(VDSO_OBJDIR)/%.o: $(VDSO_SRCDIR)/%.c
 	@$(MKDIR) $(dir $@)
 	$(call cmd,vdso_cc)
 
 $(VDSO_OBJDIR)/vdso.so: $(VDSO_OBJS)
 	$(call cmd,vdso_ld)
 
-$(VDSO_OBJDIR)/vdso-image.c: $(VDSO_OBJDIR)/vdsogen $(VDSO_OBJDIR)/vdso.so
+$(VDSO_OBJDIR)/vdso-image.c: $(VDSOGEN) $(VDSO_OBJDIR)/vdso.so
 	$(call cmd,vdsogen)
 
 $(PROG-stage3.img): linker_script $(VDSO_OBJDIR)/vdso-image.c
@@ -239,11 +236,12 @@ $(PROG-stage3.img): linker_script $(VDSO_OBJDIR)/vdso-image.c
 $(OBJDIR)/gitversion.c: $(ROOTDIR)/.git/index $(ROOTDIR)/.git/HEAD
 	$(call cmd,version)
 
-$(OBJDIR)/src/x86_64/crt0.o: $(OBJDIR)/frame.inc
+$(OBJDIR)/src/$(ARCH)/crt0.o: $(OBJDIR)/frame.inc
 
-$(OBJDIR)/frame.inc: $(SRCDIR)/x86_64/frame.h
+$(OBJDIR)/frame.inc: $(ARCHDIR)/frame.h
 	$(call cmd,sed)
 
+# XXX fix up for arm
 ifeq ($(UNAME_s),Darwin)
 ELF_TARGET=     -target x86_64-elf
 CFLAGS+=        $(ELF_TARGET)
@@ -252,5 +250,6 @@ VDSO_LDFLAGS+=  -undefined dynamic_lookup
 LD=             x86_64-elf-ld
 OBJDUMP=        x86_64-elf-objdump
 else
-LD=             ld
+# XXX already defined?
+LD=             $(CROSS_COMPILE)ld
 endif
