+++ test/e2e/nginx_1.15.6/config.json
@@ -6,5 +6,5 @@
   },
   "Boot": "../../../output/boot/boot.img",
   "Kernel": "../../../output/stage3/bin/stage3.img",
-  "Mkfs": "../../../output/mkfs/bin/mkfs"
+  "Mkfs": "../../../output/tools/bin/mkfs"
 }
