+++ test/e2e/node_v11.5.0/config.json
@@ -5,5 +5,5 @@
     },
     "Boot": "../../../output/boot/boot.img",
     "Kernel": "../../../output/stage3/bin/stage3.img",
-    "Mkfs": "../../../output/mkfs/bin/mkfs"
+    "Mkfs": "../../../output/tools/bin/mkfs"
 }
