+++ test/e2e/ruby_2.5.1/config.json
@@ -9,5 +9,5 @@
     },
     "Boot": "../../../output/boot/boot.img",
     "Kernel": "../../../output/stage3/bin/stage3.img",
-    "Mkfs": "../../../output/mkfs/bin/mkfs"
+    "Mkfs": "../../../output/tools/bin/mkfs"
 }
