+++ tools/contgen.c
