+++ tools/dump.c
