+++ tools/vdsogen.c
